rename/move class registry
