+++ src/filters/RemoveColor.ts
@@ -2,7 +2,7 @@ import { Color } from '../color/Color';
 import type { TClassProperties } from '../typedefs';
 import { BaseFilter } from './BaseFilter';
 import type { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
-import { classRegistry } from '../util/class_registry';
+import { classRegistry } from '../ClassRegistry';
 
 /**
  * Remove white filter class
