+++ src/parser/elements_parser.ts
@@ -2,7 +2,7 @@
 import { Gradient } from '../gradient/Gradient';
 import { Group } from '../shapes/Group';
 import { Image } from '../shapes/Image';
-import { classRegistry } from '../util/class_registry';
+import { classRegistry } from '../ClassRegistry';
 import {
   invertTransform,
   multiplyTransformMatrices,
