+++ src/shapes/Group.ts
@@ -19,7 +19,7 @@ import { degreesToRadians } from '../util/misc/radiansDegreesConversion';
 import { sin } from '../util/misc/sin';
 import { FabricObject, stateProperties } from './Object/FabricObject';
 import { Rect } from './Rect';
-import { classRegistry } from '../util/class_registry';
+import { classRegistry } from '../ClassRegistry';
 
 export type LayoutContextType =
   | 'initialization'
