+++ src/shapes/Path.ts
@@ -13,7 +13,7 @@ import {
   parsePath,
   type TPathSegmentsInfo,
 } from '../util/path';
-import { classRegistry } from '../util/class_registry';
+import { classRegistry } from '../ClassRegistry';
 import { FabricObject, cacheProperties } from './Object/FabricObject';
 
 export class Path extends FabricObject {
